Fixed on back pressed
Issue: For You -> PDF Reader -> (Any Bottom Tab) -> For You (PDF Reader is visible) and then press back it'll quit the app.

Solution: It navigates to For You, but doesn't recognize there's something on top of it. So we check if we can navigate up, if that's not possible then just continue w/ normal onBackPressed().